Added italian translation (#14)
Added a new it folder containing a README.md file that is a translation of the
English README.md. Added the Italian version of the license texts (convert the
html version of the Creative Commons website). Added a badge on the main
README.md pointing to the Italian translation folder. Added a badge on
it/README.md pointing to the main README.md.